+++ src/test/java/com/cobber/fta/TestBooleans.java
@@ -48,6 +48,7 @@ public class TestBooleans {
 	public void basicBoolean() throws IOException {
 		final TextAnalyzer analysis = new TextAnalyzer();
 		final String[] inputs = "false|true|TRUE|    false   |FALSE |TRUE|true|false|False|True|false|  FALSE|FALSE|true|TRUE|bogus".split("\\|");
+		final int NULL_COUNT = 2;
 		int locked = -1;
 
 		analysis.train(null);
@@ -60,14 +61,14 @@ public class TestBooleans {
 		final TextAnalysisResult result = analysis.getResult();
 
 		Assert.assertEquals(locked, -1);
-		Assert.assertEquals(result.getSampleCount(), inputs.length + 2);
+		Assert.assertEquals(result.getSampleCount(), inputs.length + NULL_COUNT);
 		Assert.assertEquals(result.getOutlierCount(), 1);
 		Assert.assertEquals(result.getMatchCount(), inputs.length - result.getOutlierCount());
-		Assert.assertEquals(result.getNullCount(), 2);
+		Assert.assertEquals(result.getNullCount(), NULL_COUNT);
 		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE +
 				"(" + analysis.getRegExp(KnownPatterns.ID.ID_BOOLEAN_TRUE_FALSE) + ")" +
 				KnownPatterns.PATTERN_WHITESPACE);
-		Assert.assertEquals(result.getConfidence(), .9375);
+		Assert.assertEquals(result.getConfidence(), 1 - (double)1/(result.getSampleCount() - NULL_COUNT));
 		Assert.assertEquals(result.getType(), PatternInfo.Type.BOOLEAN);
 		Assert.assertEquals(result.getTypeQualifier(), "TRUE_FALSE");
 		Assert.assertEquals(result.getMinLength(), 4);
@@ -85,7 +86,7 @@ public class TestBooleans {
 	}
 
 	@Test
-	public void basicBooleanYN() throws IOException {
+	public void basicBooleanYesNo() throws IOException {
 		final TextAnalyzer analysis = new TextAnalyzer();
 		final String[] inputs = "no|yes|YES|    no   |NO |YES|yes|no|No|Yes|no|  NO|NO|yes|YES|bogus".split("\\|");
 		int locked = -1;
@@ -122,6 +123,146 @@ public class TestBooleans {
 					matches++;
 		}
 		Assert.assertEquals(result.getMatchCount(), matches);
+	}
+
+	@Test
+	public void basicBooleanYesNoManySamples() throws IOException {
+		final TextAnalyzer analysis = new TextAnalyzer();
+		final String[] inputs = "no|yes|YES|    no   |NO |YES|yes|no|No|Yes|no|  NO|NO|yes|YES|no|yes|YES|    no   |NO |YES|yes|no|No|Yes|no|  NO|NO|yes|YES|bogus".split("\\|");
+		final int NULL_COUNT = 2;
+		int locked = -1;
+
+		analysis.train(null);
+		for (int i = 0; i < inputs.length; i++) {
+			if (analysis.train(inputs[i]) && locked == -1)
+				locked = i;
+		}
+		analysis.train(null);
+
+		final TextAnalysisResult result = analysis.getResult();
+
+		Assert.assertNotEquals(locked, -1);
+		Assert.assertEquals(result.getSampleCount(), inputs.length + NULL_COUNT);
+		Assert.assertEquals(result.getOutlierCount(), 1);
+		Assert.assertEquals(result.getMatchCount(), inputs.length - result.getOutlierCount());
+		Assert.assertEquals(result.getNullCount(), 2);
+		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE +
+				"(" + analysis.getRegExp(KnownPatterns.ID.ID_BOOLEAN_YES_NO) + ")" +
+				KnownPatterns.PATTERN_WHITESPACE);
+		Assert.assertEquals(result.getConfidence(), 1 - (double)1/(result.getSampleCount() - NULL_COUNT));
+		Assert.assertEquals(result.getType(), PatternInfo.Type.BOOLEAN);
+		Assert.assertEquals(result.getTypeQualifier(), "YES_NO");
+		Assert.assertEquals(result.getMinLength(), 2);
+		Assert.assertEquals(result.getMaxLength(), 9);
+		Assert.assertEquals(result.getMinValue(), "no");
+		Assert.assertEquals(result.getMaxValue(), "yes");
+		Assert.assertTrue(inputs[0].matches(result.getRegExp()));
+
+		int matches = 0;
+		for (int i = 0; i < inputs.length; i++) {
+			if (inputs[i].matches(result.getRegExp()))
+				matches++;
+		}
+		Assert.assertEquals(result.getMatchCount(), matches);
+	}
+
+	@Test
+	public void basicBooleanYN() throws IOException {
+		final TextAnalyzer analysis = new TextAnalyzer();
+		final String[] inputs = new String[] {
+				"Y", "Y", "Y", "Y", "Y", "", "Y", "", "", "Y",
+				"N", "Y", "Y", "Y", "Y", "", "", "", "Y", "",
+				"Y", "N", "N", "N", "", "Y", "Y", "Y", "Y", "N",
+				"", "", "", "", "", "", "Y", "", "Y", "Y",
+				"Y", "N", "N", "N", "", "Y", "Y", "Y", "Y", "N",
+				"Y", "", "Y", "Y", "", "Y", "Y", "Y", "Y", "",
+				"Y", "N", "N", "N", "", "Y", "Y", "Y", "Y", "N",
+				"Y"
+		};
+		int locked = -1;
+
+		analysis.train(null);
+		for (int i = 0; i < inputs.length; i++) {
+			if (analysis.train(inputs[i]) && locked == -1)
+				locked = i;
+		}
+		analysis.train(null);
+
+		final TextAnalysisResult result = analysis.getResult();
+
+		Assert.assertNotEquals(locked, -1);
+		Assert.assertEquals(result.getSampleCount(), inputs.length + 2);
+		Assert.assertEquals(result.getOutlierCount(), 0);
+		Assert.assertEquals(result.getMatchCount(), inputs.length - result.getOutlierCount() - result.getBlankCount());
+		Assert.assertEquals(result.getNullCount(), 2);
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_BOOLEAN_Y_N));
+		Assert.assertEquals(result.getConfidence(), 1.0);
+		Assert.assertEquals(result.getType(), PatternInfo.Type.BOOLEAN);
+		Assert.assertEquals(result.getTypeQualifier(), "Y_N");
+		Assert.assertEquals(result.getMinLength(), 1);
+		Assert.assertEquals(result.getMaxLength(), 1);
+		Assert.assertEquals(result.getMinValue(), "n");
+		Assert.assertEquals(result.getMaxValue(), "y");
+		Assert.assertTrue(inputs[0].matches(result.getRegExp()));
+
+		int matches = 0;
+		for (int i = 0; i < inputs.length; i++) {
+			if (inputs[i].trim().matches(result.getRegExp()))
+					matches++;
+		}
+		Assert.assertEquals(result.getMatchCount(), matches);
+	}
+
+	@Test
+	public void basicBooleanN_Bad() throws IOException {
+		final TextAnalyzer analysis = new TextAnalyzer();
+		final int countN = 30;
+		final int countC = 10;
+
+		for (int i = 0; i < countN; i++)
+			analysis.train("N");
+		for (int i = 0; i < countC; i++)
+			analysis.train("C");
+
+		final TextAnalysisResult result = analysis.getResult();
+
+		Assert.assertEquals(result.getSampleCount(), countN + countC);
+		Assert.assertEquals(result.getOutlierCount(), 0);
+		Assert.assertEquals(result.getMatchCount(), countN + countC);
+		Assert.assertEquals(result.getNullCount(), 0);
+		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_ALPHA);
+		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
+		Assert.assertNull(result.getTypeQualifier());
+		Assert.assertEquals(result.getMinLength(), 1);
+		Assert.assertEquals(result.getMaxLength(), 1);
+		Assert.assertEquals(result.getMinValue(), "C");
+		Assert.assertEquals(result.getMaxValue(), "N");
+		Assert.assertEquals(result.getConfidence(), 1.0);
+	}
+
+	@Test
+	public void justY() throws IOException {
+		final TextAnalyzer analysis = new TextAnalyzer();
+		final int COUNT = 50;
+
+		for (int i = 0; i < 50; i++)
+			analysis.train("y");
+
+		final TextAnalysisResult result = analysis.getResult();
+
+		Assert.assertEquals(result.getSampleCount(), COUNT);
+		Assert.assertEquals(result.getOutlierCount(), 0);
+		Assert.assertEquals(result.getMatchCount(), COUNT);
+		Assert.assertEquals(result.getNullCount(), 0);
+		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_ALPHA);
+		Assert.assertEquals(result.getConfidence(), 1.0);
+		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
+		Assert.assertNull(result.getTypeQualifier());
+		Assert.assertEquals(result.getMinLength(), 1);
+		Assert.assertEquals(result.getMaxLength(), 1);
+		Assert.assertEquals(result.getMinValue(), "y");
+		Assert.assertEquals(result.getMaxValue(), "y");
+		Assert.assertTrue("y".matches(result.getRegExp()));
 	}
 
 	@Test